Refactor: Use Prisma's typedSql and LibSQL adapter
This commit updates the project to use Prisma's typedSql feature for more type-safe SQL queries. It also integrates the LibSQL adapter, enabling the use of LibSQL as the database. This includes changes to Prisma schema, route loaders, and dependency management.
