+++ eslint.config.js
@@ -9,6 +9,6 @@ export default [
 		rules: { 'react-hooks/rules-of-hooks': 'off' },
 	},
 	{
-		ignores: ['.react-router/*', './app/utils/prisma-generated.server'],
+		ignores: ['.react-router/*', './app/utils/prisma-generated.server/**'],
 	},
 ]
