--- prisma.config.ts.bak
@@ -1,18 +1,0 @@
-import * as path from 'node:path'
-import { PrismaBetterSQLite3 } from '@prisma/adapter-better-sqlite3'
-
-// Import environment variables
-import 'dotenv/config'
-
-export default {
-	schema: path.join('prisma', 'schema.prisma'),
-	async adapter() {
-		const url = process.env.DATABASE_URL
-		if (!url) {
-			throw new Error(
-				'DATABASE_URL is required (expected a file: URL for SQLite).',
-			)
-		}
-		return new PrismaBetterSQLite3({ url })
-	},
-}
