+++ prisma/schema.prisma
@@ -2,7 +2,7 @@
 // learn more about it in the docs: https://pris.ly/d/prisma-schema
 
 generator client {
-  provider  = "prisma-client-js"
+  provider  = "prisma-client"
   engineType = "library"
   output    = "../app/utils/prisma-generated.server"
   previewFeatures = ["typedSql"]
